Further clean-ups to ASDF system definitions.
Use standard form of "cl-async/foo" instead of "cl-async-foo" for
component names.

Use `:pathname` option to make the components tidier.

This also fixes the issues with the reloading warning, which I believe
happened because ASDF was rereading the ASD files because of the
non-standard system names (although I might be wrong about this
hypothesis).